# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/regions": {
      "get": {
        "description": "Returns all sales regions for assignment and reporting.",
        "operationId": "AdminRegion_listRegions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of regions"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List regions (supervisor / admin)",
        "tags": ["Admin regions (supervisor / admin)"]
      },
      "post": {
        "description": "Creates a region. Slug is generated from `name` if omitted (unique suffix added on collision). User.regionId must match a region id (cuid), not the slug.",
        "operationId": "AdminRegion_createRegion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Nairobi West"
                  },
                  "slug": {
                    "type": "string",
                    "example": "nairobi-west",
                    "description": "Optional; otherwise derived from name"
                  },
                  "isActive": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Region created"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          },
          "409": {
            "description": "Slug already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create region (supervisor / admin)",
        "tags": ["Admin regions (supervisor / admin)"]
      }
    },
    "/admin/regions/{id}": {
      "patch": {
        "description": "Partial update (slug, name, isActive).",
        "operationId": "AdminRegion_updateRegion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Region id (cuid)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Region updated"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          },
          "404": {
            "description": "Region id not found"
          },
          "409": {
            "description": "Slug already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update region (supervisor / admin)",
        "tags": ["Admin regions (supervisor / admin)"]
      }
    },
...
  },
  "info": {
    "title": "Engaged Sales API",
    "description": "API documentation for Engaged Sales backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}